refactor: major project restructuring and improvements
BREAKING CHANGES:
- Reorganized project structure for better maintainability
- Moved scripts to categorized folders (build/, test/, debug/, setup/)
- Consolidated test files in tests/integration/
- Moved documentation to docs/ with phases/ subfolder

Major changes:
- Replaced index.ts with refactored version
- Updated all import paths in test files
- Created organized directory structure
- Added comprehensive .gitignore
- Updated package.json scripts for new paths
- Legacy code moved to legacy/ folder

Improvements:
- Better separation of concerns
- Cleaner project root
- More intuitive file organization
- Easier navigation and maintenance
