+++ .gitignore
@@ -1,55 +1,58 @@
 # Dependencies
 node_modules/
+package-lock.json
 
 # Build output
 dist/
 build/
-
-# TypeScript cache
 *.tsbuildinfo
 
+# Test files
+test-*.txt
+*.test.txt
+test-output/
+test-extract/
+*.enc
+
+# Archives
+*.zip
+*.tar
+*.tar.gz
+
+# IDE
+.vscode/
+.idea/
+*.swp
+*.swo
+*~
+
+# OS
+.DS_Store
+Thumbs.db
+
 # Logs
-logs/
 *.log
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-# Runtime data
-pids/
-*.pid
-*.seed
-*.pid.lock
-
-# Directory for instrumented libs generated by jscoverage/JSCover
-lib-cov/
-
-# Coverage directory used by tools like istanbul
-coverage/
-*.lcov
-
-# nyc test coverage
-.nyc_output/
-
-# Editor directories and files
-.idea/
-.vscode/
-*.swp
-*.swo
-*~
-
-# OS files
-.DS_Store
-Thumbs.db
-
-# Environment variables
+# Environment
 .env
 .env.local
 .env.*.local
 
-# Backup files
-*.backup.*
+# Coverage
+coverage/
+.nyc_output/
 
 # Temporary files
 tmp/
 temp/
+*.tmp
+*.temp
+
+# Backup files
+*.backup
+*.bak
+backup-*/
+*-backup-*/
