--- build-project.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-cd /Users/sangbinna/ai-filesystem-mcp
-npm run build
