--- build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-cd /Users/Sangbinna/mcp/ai-filesystem-mcp
-npm run build
