+++ claude-config-example.json
@@ -3,7 +3,7 @@
     "ai-filesystem": {
       "command": "node",
       "args": [
-        "/Users/Sangbinna/mcp/ai-filesystem-mcp/dist/index-refactored.js"
+        "/Users/Sangbinna/mcp/ai-filesystem-mcp/dist/index.js"
       ]
     }
   }
