+++ docs/CLEANUP_REPORT.md
@@ -1,0 +1,30 @@
+# 프로젝트 정리 보고서
+
+## 정리 완료 항목
+
+### 📁 디렉토리 구조
+- `docs/` - 모든 문서 파일 정리
+  - `phases/` - PHASE 관련 문서들
+- `scripts/` - 스크립트 파일 카테고리별 정리
+  - `build/` - 빌드 관련 스크립트
+  - `test/` - 테스트 실행 스크립트
+  - `debug/` - 디버그 도구
+  - `setup/` - 설치 및 설정 스크립트
+- `tests/` - 테스트 파일 정리
+  - `integration/` - 통합 테스트
+- `legacy/` - 리팩토링 이전 코드
+
+### 🔄 주요 변경사항
+1. `src/index.ts` - 리팩토링된 버전으로 교체
+2. 모든 테스트 파일을 `tests/integration/`으로 이동
+3. 스크립트 파일들을 목적별로 분류
+
+### ⚠️ 필요한 후속 작업
+1. `package.json`의 스크립트 경로 업데이트 (update-package-json.js 실행)
+2. 각 스크립트 파일 내부의 상대 경로 확인 및 수정
+3. CI/CD 설정 파일 경로 업데이트 (있는 경우)
+4. README.md의 파일 경로 업데이트
+
+### 🗑️ 삭제 가능 항목 (확인 후)
+- `legacy/` 폴더 (모든 기능이 정상 작동 확인 후)
+- 백업 폴더 (충분한 테스트 후)
