+++ package.json
@@ -6,25 +6,46 @@
   "type": "module",
   "scripts": {
     "build": "tsc",
+    "build:check": "./scripts/build/check-build.sh",
+    "build:diagnose": "./scripts/build/diagnose-build.sh",
+    "build:quick": "./scripts/build/quick-build.sh",
     "build:watch": "tsc -w",
-    "start": "node dist/index.js",
-    "start:refactored": "node dist/index-refactored.js",
+    "clean": "rm -rf dist",
+    "debug:failed": "node scripts/debug/debug-failed.js",
+    "debug:registry": "node scripts/debug/debug-registry.js",
+    "debug:step": "node scripts/debug/debug-step-by-step.js",
     "dev": "tsx watch src/index.ts",
     "dev:refactored": "tsx watch src/index-refactored.ts",
-    "test": "npm run build && node test.js",
-    "test:refactored": "npm run build && node test-refactored.js",
-    "test:jest": "jest",
-    "test:unit": "jest --testPathPattern=unit",
-    "test:integration": "jest --testPathPattern=integration", 
-    "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage",
-    "test:phase1": "npm run build && node test-phase1.js",
-    "validate:phase1": "npm run build && node debug-phase1.js",
-    "clean": "rm -rf dist",
+    "diagnose": "node scripts/debug/quick-diagnose.js",
+    "format": "prettier --write 'src/**/*.ts'",
     "lint": "eslint src --ext .ts",
-    "format": "prettier --write 'src/**/*.ts'"
+    "setup": "./scripts/setup/install.sh",
+    "setup:clean": "./scripts/setup/clean-install.sh",
+    "setup:validate": "./scripts/setup/setup-and-validate.sh",
+    "start": "node dist/index.js",
+    "start:refactored": "node dist/index-refactored.js",
+    "test": "npm run build && node tests/integration/test.js",
+    "test:all": "npm run build && node tests/integration/test-all-39.js",
+    "test:coverage": "jest --coverage",
+    "test:git": "npm run build && node tests/integration/test-git.js",
+    "test:integration": "jest --testPathPattern=integration",
+    "test:jest": "jest",
+    "test:metadata": "npm run build && node tests/integration/test-metadata.js",
+    "test:phase1": "npm run build && node tests/integration/test-phase1.js",
+    "test:refactored": "npm run build && node tests/integration/test-refactored.js",
+    "test:transaction": "npm run build && node tests/integration/test-transaction-deep.js",
+    "test:unit": "jest --testPathPattern=unit",
+    "test:watch": "jest --watch",
+    "validate:phase1": "npm run build && node scripts/debug/debug-phase1.js"
   },
-  "keywords": ["mcp", "filesystem", "ai", "code-manipulation", "refactoring", "security"],
+  "keywords": [
+    "mcp",
+    "filesystem",
+    "ai",
+    "code-manipulation",
+    "refactoring",
+    "security"
+  ],
   "author": "AI FileSystem Team",
   "license": "MIT",
   "dependencies": {
