+++ scripts/debug/check-build-error.js
