+++ scripts/debug/debug-registry.js
