+++ scripts/debug/quick-diagnose.js
