+++ scripts/setup/check-path.sh
