+++ scripts/setup/clean-install.sh
