+++ scripts/setup/install.sh
