+++ scripts/setup/setup-and-validate.sh
