+++ scripts/test/rebuild-and-test.sh
