+++ tests/integration/test-metadata.js
